Fix automatic inheritance.
It is now possible to subclass any nested class or module in a class inheriting from Report (like CostQuery) at the same time keeping it optional. That way no code has to be copy-pasted from the engine and a hook system is largely uneccessary. No alias_method_chain needed, as it works essentially like having Module#prepend with the notable exception of being able to run two seperate reporting subclasses in the same process. This is necessary for the Usage Reporing (creating reports both on logins and courses, for example) and might come in handy if we want to create reports on issues or other entities in Redmine.

Finn: #3416, Cockpit: #18043.
Signed-off-by: Konstantin Haase <konstantin.mailinglists@googlemail.com>