Meeting show refinements. #2871
Changed the context menu generation to a helper for more flexibility, the visibility of elements has been changed to use show-meeting_agenda/show-meeting_minutes and edit-meeting_agenda/edit-meeting_minutes classes in a post-loaded js snippet. This relieves from having to set the visibility of each element individually. The sole exception is the edit pane itself for the content because hiding it post-load would flicker too much.