Avoid calling attr_protected if _accessible is used
When attr_accessible is used, we may assume that watcher_ids and
watcher_user_ids is already protected from mass assignment. Therefore we do not
need to perform additional actions.

Also, when attr_protected and attr_accessible is used on the same model, AR
throws nasty "Declare either attr_protected or attr_accessible for *, but not
both." errors.