Fixing inherit_version_to_leaf_tasks after core behavior changed
Previously save was called whether or not an issue was actually changed. That
way we could be sure, that updates to the fixed_version_id tickled up the
issues tree. This is no longer valid. Therefore we need to update all
descendants whenever our fixed_version_id changes. This should happen for
version changes in stories and if tasks are moved to other stories.