Perf opt - touching journaled w/o callbacks
Initially the journaled was only touched on journal creation to make sure, that
journal updates, that do not update the journaled element itself (like adding a
note) cause a change in updated_(on|at). This has gone a bit overboard, since
simply touching the journaled causes a roundtrip to the db and triggering of all
save callbacks on journaled in all cases.

These changes try to avoid the db roundtrip by only saving when things actually
changed and avoid triggering the callbacks by using update_all instead of save.