Making test more generic
I don't know why the behavior changed with the last commit, but [3,4] has now
become [1,2]. Therefore, I'm reducing the specifity of this test, to only check
if parent has no children. Also, since the root is the issue itself, the default
lft and rgt values should have been 1, 2 anyway and the tested behaviour
seems to have been wrong in the first place. This seems to be confirmed by
the fact, that redmine tests for [1,2] as well (see
http://www.redmine.org/projects/redmine/repository/revisions/6311/diff/trunk/test/unit/issue_nested_set_test.rb)