fixes issue diff link when having a relative_url_root defined
Because relative_url_root gets applied along with Setting.host_name having the same root when constructing the link the root is included two times in the resulting link. This is accounted for with this fix. There is a better method of doing so but this would require OpenProject to use rails 3.2 and we are not there yet.