fixes issues history overlapping issues changeset
Formerly the history of an issue would overlap the issue's changeset which would result in the user not beeing able to click on the links in the changeset. The solution is not perfect. It would look much nicer if the changesets would float right. But this is not possible with the content itself having no fixed with.