renders the action menu twice on issue#show instead of copying it with javascript
The js was only used on this specific page and it was required in the body. Now we render the action menu twice. Doing so required to place it in a partial which of course has some performance overhead. This and the fact that the action menu does not follow the paradigms employed by the other menus will require us to refactor the action menu to be just another menu, albeit it being rendered differently.