allows joining multiple tables of awesome_nested_set structured models while using nested_set methods
When joining issues and projects both tables come with their set of nested_set attributes. The attributes are named equally in both tables. This leads to ambiguous references when calling e.g. Issue.find(1).visible.descendants.

This fix scopes the nested_set to reference the left attribute fully qualified.

Adds additional methods that can be applied should the need arise.

We should reevaluate why nested set is vendored as the problem would not have arisen when using the gem.