eager loads models associated with the issues ancestors and descendants
Additionally it fixes the bug that all issues of the tree would be displayed even when the user didn't have the right to view them. This would be the case when the user has rights in Project A but non in Project B. Viewing an issue in Project A having ancestors/descendants in Project B would show those erroneously.

Saves about 1.5 seconds when having 100 ancestors/desendants.