uses ajaxAppend for loading the issue#edit form
This means that it is no longer necessary to render the edit form on
every show request which will significantely increase the load speed
of the page. We will have to see if it is ok for the user to have to
wait for the form to load once he actually wishes to edit the issue.