include calendar js files in case they are needed for edit
The edit form requires Calendar.js. As the form is loaded asynchronously now and thus without html-header, the required files where no longer included in the page once an edit is desired. This might not be an optimal solution as the files are loaded regardless of whether they are actually used or not. Given the migration to rails 3, I consider it a practical solution as it solves the problem without creating mechanisms that might later be on hand anyway once the migration is done.