completely removes custom allowance evaluation (granular_for, inheritance)
This should make the code easier to understand and more performant.

As part of this commit a lot of falsely presented information is no longer shown to users who have limited rights. E.g. a user having only the view_own_hourly rates sees only the planned and actual labor costs for himself when viewing a budget.