refactors participants as watchers from meeting_participants to meeting
As the meeting_participants model in my opinion should be hidden as much as possible and as such have the character of just a data object, the behaviour of adding all participants as watchers is better placed in meetings.

Doing this has the added advantage, that the author is just another participant and as such is added to watchers along with the rest.

The consequence of doing this is, that creating a meeting_participant directly will no longer lead to the participant beeing a watcher.