Attempt to fix custom_fields cukes
Step definitions shouldn't react on page state, but always execute the
same actions. Otherwise they need to wait a long time, only then they
can be sure their decision is right. Normally, capybara waits for a
specific event and only waits the full time when the event never
occurs.

Changed the 'I add the (.+) localization of the "(.+)" attribute as
"(.+)"' step to only add localization attributes. It fails most of the
time when used for changing an attribute. Adapted all cukes found
failing through this.