prevents attribute name to be humanized
Without the :default => false, the method's super implementation will
use attribute_name.to_s.humanize as a last resort. This will prevent an
error from beeing raised which will render the rescue statement added
for backwards compatibility useless.

Adding the fix will lead to a lot of warnings. But hey, that will lead
to a faster adaptation of the new standard.