removes ambiguity of plugin cucumber steps with core
Removes the ambiguity of the plugin's cucumber steps with those from the
core when executing cucumber by the core's rake task. Doing so loads all
the step definition leading to the mentioned ambiguity.

This also tries to elevate the scenario description away from web steps
to steps having a specification character.