Introduces strong_parameters whithout making use of it yet
This introduces the strong_parameters gem. It will be default
in rails4 so I consider it a good idea to start using it now.

This commit does not yet make use of the possibilities.
It is intended so we start using it and in order for it to be
available to plugins.