ensures cleaning up User.current after each step
When User.current is set by a spec, an instance variable of the class
gets set to the provided value. This instance variable is not reset
between specs to the next spec will "profit" as well.

This commit implements cleaning up after such lazy specs.

The lazy specs should be improved so that this cleanup is no longer
required. However I do not have the time to do this right now. As a
reminder I implemented a very verbose warning which should give us
enough of an encouragement that we will improve the specs eventually.