introduces ajax triggered inline diff on work_package#show
The functionality itself is adapted from issue#show.

On the other hand, the diff is no longer rendered with the full url as the
link's href. Using the full url caused problems with selenium tests that
ran into problems with cross domain requests. This could have been fixed
by setting the application's host and protocol setting accordingly.
However, it is much cleaner to use a relative url. So for rendering the
diff inside the application the behaviour was alterd to use relative
url. The full url is now only used where it is absolutely necessary
(pdf, email).