Improve cucumber task startup time
Allow running cucumber in the same process as Rake. This saves about
10 seconds on my machine for loading the Rails environment a second
time.

Setting RAILS_ENV=test is required, so Bundler loads Gems from the
test group. This happens in  application.rb and before the Rake tasks
are loaded. So setting the environment in the Rake task would be too
late.