Add environment-specific seeds and load plugin-seeds
The commit adds a mechanism to load seeds that are specific for the
given environment. The seeds are stored in db/seeds/<environment>.rb
and should be written idempotent, to make sure, they don't
destroy/overwrite anything when being run multiple times.

The seeding now also loads the seeds for the plugins, which allows
plugins to add plugin-specific constants to the database.