Added a sample project to the development-seed.[skip ci]
using RAILS_ENV=development rake db:seed, you can now generates a sample
project: The project contains several issues, a timeline and lots of
planning-elements to play with. Feel free to add some more sample data
to this pet-project.

The project is re-generated every time you seed the database, so every
changes to this project will be lost.

The ci is currently skipped, as it currently does not run the seeding.