Adds proper limits on wiki content fields.
The limits do not apply to postgres, which is why the schema.rb never
included them. MySQL would have produced a different schema.rb.
Subsequent migrations would break when the limit was not established
in MySQL, leading to a much shorter limit. Yes, that is true, no limit
results in a field of shorter size than the maximum size for fields in
MySQL. Hooray for smart defaults!