moves #duration and #reschedule_after from subclasses to work_package
This changes the default behavior as it was for issues so that the duration calculation includes start and due dates. As an example, a work package from today to today has a duration of 1. 1 is also returned if start or due date is not set.

Reschedule after is moved unchanged, even if there might be some issues and inconsistencies. E.g. it depends on the "update_parent_attributes" method to propagate the due/start. This leads to an inconsistency where a reload is required if the wp is not a leaf while no reload is required when it is.