Half-way fixed the cuke to work again.
Unfortunately, this is only half the deal: the blocks-names are also
used as keys. These keys are not only stored in the db (to store the
configuration of the my_project_page), but also used to lookup labels
for the i18n. To make matters even worse, there is already a migration,
that migrates the old, twisted issue-keys (issueassignedtome) into
proper issue-key (issues_assigned_to_me). In order to really, really get
rid of issue we need to:

  * rename the templates
  * rename the labels in de/en.yml
  * rename the keys in the migration

which properly blows up the scope of this fix. I therefore close this
fix here halfway, and we need to open up a new ticket to get rid of
issue properly.