Added a special handling for all user-filters (responsible_id, assignee_id)
There was previously a list of all user-fields, that was used before to
add soome speciall operations (e.g. replacing <<mee>> with the id of the
currently logged in user) - this list has now been replaced with an
attribute on the filter (:user_filter => true) - this allows to add
special handling for these kind of fields at different parts in query.

The query-spec is inherintly broken: We currently do not really have
any unit-specs for query, only the tests and the (integration) api-feature.

We probably should migrate the tests to proper specs, but as we should
also refactor query (and check that we have decent coverage for all
query-cases) - this should probably be done together.