replace the rendering of planning elements with rabl
added oj and multijson, as they are supposedly faster as oj is
supposedly faster than the default-engine: There are still lots of
switches in rabl left to improve the speed even more.

Currently, only the inde-templates have been replaced, other templates
have been left untouched: The tests now break, as there is a special
test for the replaced .rsb-file: This is a very specialised test, we
should probably replace this with tests for an expected json-response.