Fixes a scrolling problem w/ IE10.
When a timeline collapses after a modal is closed, IE10 would sometimes
still display a part of the page where there is no content. This happens
when work packages are moved around and nested so that the resulting
timeline is shorter than the previous scrollTop, or when the timeline
falls back to the default timeline expansion (a bug currently not
fixed).