removed dependent objects and using pure structs in the view
to solve our performance-issues, this commit introuces two changes:

- removing all dependent objects from the view
- using structs instead of ar

the first change reduces the number of objects to render, the second
change makes sure, that we only hit the db once and makes really, really
sure that we don't have any n+1-queries.

This breaks a lot of tests for the formats, which will be fixed in the
following commits.