Fixes auto update text reference migration
This commit just changes the order of the regexes applied to a text. So,
the automatic update text 'Updated automatically by changing...' is
migrated first. This is required because the general regex does not
respect the specifics the the automatic update text (e.g., the trailing
underscore in '...values within child planning element *123_').