Fixes a bug with the timelines loader.
Timelines loading relies on project filtering before fetching planning
elements. It also relies on fetching users based on only the required
ids from projects and planning elements.

This broke when filtering on responsibles of projects, since the filter
on users happened before users were loaded. Now, responsible filtering
of projects has a fallback on filtering based on id only, so the user
needs not be fetched.