Fix rake cucumber:plugins when being run with RAILS_ENV=test
The rake task previously gave empty arguments to cucumber, which
cucumber doesn't like.

The following is my theory on why this works when the rake task is
started in a non-test environment. In the test environment, cucumber
is set to not fork for startup performance. When forking, parameters
are given to cucumber in another way (as command line parameters) and
possibly processed differently. This results in the empty arguments
being skipped, while when not forking these empty arguments are
treated as feature folders that can't be found obviously.

With this fix the empty arguments are not given to cucumber.