Omniauth: Fix user redirect after pending account and login
When registering an account, the account was pending and the user
logged in from the same login page after activation, the user was
redirected to the authentication provider (at least that's what the
back_url was being set to, this would probably have been blocked later
by redirect_back_or_default). Setting the back_url explicitly has the
advantage of redirecting the user to the original page where the user
clicked on the Omniauth link.