Make 'should see' step definition case-insensitive
Many Cucumber features were broken as a result of the styling change
introduced in fc473559. Cukes should test behaviour and semantics,
_not styling_. `text-transform: uppercase` does not convey semantic
meaning in the context of these particular headings.

Selenium will match text as it is rendered and visible to the
end-user. Thus, by default, we need to make the text matching in the
"should (not) see" step definitions less strict.

Signed-off-by: Alex Coles <alex@alexbcoles.com>