WIP started wiring up option for shown in all projects.
Pretty interesting problem here in that shown in all projects means that the query has a nil project. However when viewing work packages the query's project gets set to the current project, which means that given the current setup it's impossible to know whether the query is to be shown for all projects or not. In order to implement this I think we'll have to add a new field to the query and then change the backend logic which is a massive pain in the arse.