Omniauth: Add after_login, remove on_success
After a successfull login seems to be a much clearer description of
when the callback will be executed than "on successful authorization".
E.g. it's not clear whether ending up on the registration page also
indicates successful authorization (which it does in the OmniAuth
sense, but not in the OpenProject sense).

The after login callback is called each time a user is logged in, no
matter whether that happens via regular Omniauth login or after
completing the registration process and being automatically logged in
(when automated account activation is enabled).