Make WorkPackagesAPI PATCH expect JSON request body
Parameters not included in the URL should be encoded as
`application/json` (rather than `x-www-form-urlencoded`).

This patch causes the raw JSON request body to be passed through to
the WorkPackageRepresenter's `#from_json` method, rather than letting
Grape deserialize (with its built-in "parsers") the data and then
having to reserialize the parameters.

Signed-off-by: Alex Coles <alex@alexbcoles.com>