removes unnecessary visibility filter on Version scope
Authorization is handled an project level here and thus does not need to
be done for the version itself. Authorising on the Version in addition
has the drawback to only return versions in projects which the user can
see work packages in. It does not take into account that the version can
be shared.