API v2 PE index: Allow disabling parent rewiring
Allow disabling rewiring - this exposes parent IDs of work packages invisible
to the user.
When requesting single work packages via IDs, the rewiring fails as it assumes
that all visible work packages are loaded, which they might not be. For a work
package with a parent visible to the user, but not included in the requested IDs,
the parent_id would thus be nil.
Disabling rewiring allows fetching work packages with their parent_ids
even when the parents are not included in the list of requested work packages.

There's also a case when rewiring is not used at all: When not providing a project
id and requesting via /api/v2/planning_elements.json?ids=[...]. In this case
parent_ids were and are always shown. This is pretty inconsistent, but that's
the way API v2 is. Btw, planning_elements.json returns 404 when not giving an ids
parameter.

Apparently, the way custom fields are returned is also different between
the general planning elements endpoint and the project-specific one. I didn't
look into this, though.