WP Index: Fix total sums overlaying WPs in Firefox
cg-busy adds a 'position: relative' to the CSS which makes Firefox
position the absolutely positioned div with the sums relative to the
total sum table row, which has height 0 due to no content (the div is
ignored for 'position: absolute').

Chrome seems to ignore the 'position: relative'.

Removing cg-busy for the total sums as the loading indicator doesn't
seem to be shown anyway.