suppresses action dispatch routing errors
The errors:
* AbstractController::ActionNotFound
* ActionController::RoutingError
are no longer being logged and are displayed as an error native to the
application.

This prevents bots looking for weaknesses from causing huge amounts of
errors which might trigger application monitoring to raise alarms.

Additionally, the error pages displayed to genuine users can be
presented as part of the application.