TimeEntries spent_on date display fix.
The detail view of logged time entries does not show the logged date
(spent_on), but instead displays the current time.

I believe this is an error in the data passing in the angular scope
which can be fixed by adding a data observer to the date attribute.

This commit appears to fix the issue.