Fix WP validation
In contrast to my expectation the contract does not trigger the WP's validation.
Thus, those validations need to be triggered separately.

This issue became visible when removing the call to WP's valid? method removed.
The call is removed because it empties the list of errors which removes those
errors added by the contract.