Move WP move test to spec and update its specification
With the additional validation of a WP's children, a move of a WP to another
project is only valid if cross project relations are allowed. The old test
didn't consider that setting and just assumed that a WP move is valid across
project borders in any case (even if cross project relations are forbidden).