Revert "Bring in legacy styles with require directive"
The change is in conflict with the themes.

As themes import the default.css.sass again in order to have their
styles applied to the rules, the complete css should be handled by saas.

Otherwise, files handled via sprockets are not included as an include
will not contain "require" files.

This reverts commit 240c7eeb884afba35a34fa7f04cf0896acfb4605.

Conflicts:
	app/assets/stylesheets/default.css.sass