Extend apiv3 blueprint for user (un-)locking
With OmniAuth, we now have external services authenticate
users in an asnychronous manner. With this change, proper means
to manage the status based from these services become a necessity.
Particularly, deleting and locking users based on outer events are
a requirement for timely account administration.

As the API(v3) already provides the functionality to remove user
accounts, this commit suggests an extension to provide locking
and unlocking user accounts through the API.