Allow bypassing basic auth for repository access authorization
In our use-case, we use RADIUS to authorize users through Apache
in combination with an OmniAuth strategy.

Thus, we do not have an active authentication source to
perform basic auth against.

This commit adds an optional setting to allow bypassing
basic auth in the SysController.

When authorizing the user directly in the web server,
passing the logged in user to SysController is sufficient
to retrieve the project permissions.

The relevant work package:
https://community.openproject.org/work_packages/18356