fix controller response
As the diff action claims to be more generic, react with a 404 instead
of a 403 whenever the journable type is not a WorkPackage.

This makes the the use of the diff action less generic, as a user will
still need the view_work_pacakges permission to access the diff action.