fix features for work packages
The fix consists mostly of rewriting the test for a missing subject into
a test for a subject that is too long.

This is due to the fact that with the new form builder, the forms make
use of the `required` attribute for form fields in HTML5 documents. This
prevents the form to be sent w/ an empty, but required field.