Do not use npm-shrinkwrap.json when packaging npm dependencies.
This avoids downloading a huge amount of things that are only needed in
development. It will reduce the package size, and certainly improve the
reliability of the packaging process (i.e. less ECONNREFUSED when
downloading phantomjs and such).

This commit also updates the NodeJS buildpack, and forces the use of the
HTTP npm registry for better performance and reliability when packaging.