attempt to stabilize query menu item cuke
The cuke used to press on the "Settings" button regardless of whether
the page was fully loaded. In cases the page was not loaded, this
resulted in a no-op as the "Settings"-menu will only open when the page
has finished loading. Therefore steps after the click that assumed the
menu to be open failed.