"fix" too much space before content in Wiki
This is a dirty hack to achieve the goal. Nobody should ever have to do this in his live, but there are deadlines and issues considered important...

This fixes the issue the dirty way, but it works and has few implications on all other parts of the application.

I hope someone will come and delete those lines quickly...