Replace UserMailer plain text hack w/ Interceptor
The previous hack was broken, most likely as a result of internal
 `ActionMailer::Base` changes.

This solution is still not ideal as the resulting email is
unnecessarily multipart: albeit `multipart/mixed` with one plain text
part, rather than `multipart/alternative`. There does not appear to
be a built-in Ruby Mail API to convert multipart back to plain text.

Signed-off-by: Alex Coles <alex@alexbcoles.com>