reintroduce atwho rules
- let them live in their own file instead of _misc_legacy.sass
- use colors given by theme

I introduced global variables for font colors that appear on a given color variable as background,
I feel this is neccessary to be able to use the primary-color etc. as background and still ensure good accessibility (e.g. if a theme choses to use a light primary color this would make white a bad choice for a font)