extend allow_attribute_update with action
As we now have not only an update action but also create actions we also
have two types of contracts (Create and update). As the attributes that
are allowed to be changed might differ between the two, we enable a
plugin to specify for which action it wants to extend the updatable
attributes.