Repository Management - Refactoring and Preparation
Related work package: [#20415](https://community.openproject.org/work_packages/20415).

This is a combined refactoring and extraction effort of local SCM functionality for SVM, Git adapters in preparation for the explicit lifetime management of repositories in #20416 and implementation of (local) repository storage quota in #18393.

This PR is still heavily WIP and for the time being, depends on the **legacy specs** for testing.
To run the tests, execute the following steps:

~~~
rake test:scm:setup:all
rake spec:legacy
~~~

I'm open for ideas on how we integrate scm adapters tests into current specs.

---

- [x] Exchange `IO::Open` for `Open3` to avoid countless manual exitcode checks
- [x] Extract local functionality in its own Module
  - [] Extend `LocalClient` module with links to create/delete services for #20415
- [x] Break down unbearable local parse routines (Hound will be over the place until fixed - but down from 260 to < 20)
  - [x] `entries`
  - [x] `revisions`
- [] local repository factory
- [] Specs
  - [] Fix / Modernize
  - [] More adapter-related specs