Fix broken routes to Cost report due to url_for namespaces
A spec concerned with that problem is `spec/features/menu_spec.rb`.
Within it, the correct solution is noted, but not applied to the menu items.

Depending on the current page the link to the cost reports was broken.
This seems to be due to a peculiarity of the rails routing where
`url_for controller: :foo` would return a link relative to the controller
handling the current request path if the controller was routed to via a
namespaced route.

Example:

`url_for controller: 'cost_reports'` will yield different results ...

when on `/projects/ponyo/work_packages`: `/projects/ponyo/cost_reports` (correct)
when on `/projects/ponyo/work_packages/calendar`: `/work_packages/cost_reports?project_id=ponyo`

This is only relevant for project menu entries, not global ones (`project_id` param is nil)*.
Meaning that you have to make sure to force the absolute URL in a project menu entry
by specificying the controller as e.g. '/cost_reports' instead of just 'cost_reports'.