Properly target breadcrumbs entries.
The current features/menu_spec.rb selects Cost reports from the entire
.breadcrumb using a `text:` selector.

As the tests are working locally, I can only assume this is connected to
how capybara evaluates the text selector.
If it selects text using innerText, this may fail for some browsers.

The spec is now addressed on the breadcrumb li item diirectly.