Frontend extensions and build refactoring
- switched to toggleable-attribute-groups to provide means for
adding buttons in the attribute header (e.g., delete button).

- added separate destroy_info page with different notifications
for managed and linked repositories to ensure ramifications of removing
physical repositories are communicated to the user.

- Fixed several issues regarding scm_type selection and the building
  of repositories.

Open questions still concern the question of acessing unsupported
repositories (e.g., the soon-to-be-removed filesystem adapter).
Acessing the class should will then throw an error, but we still have
to allow users to remove the affected repository (somehow).