Remove filesystem adapter
This commit removes the entire FileSystem repository and adapter.
It rids all specs (core and legacy) of remaining uses of this adpater
and replaces it with a temporary subversion repository, where necessary.

We also need to remove remaining filesystem instances from the database,
otherwise users will receive errors when accessing said repositories
on the Frontend.

This does not destroy any data on filesystem, but database information
on the repositories (remarks) are lost.