Fix code syntax highlighting output.
When annotating a file in the repository for whose language CodeRay supports,
the output is not properly rendered as HTML, but escaped.

As CodeRay escapes any items for HTML before output, we can safely
assign syntax lines to HTML when highlighted.
If text is output from CodeRay (no language set), we want to escape it
instead.