[ci skip] Allow notification-box.warning to be severe without being an error.
In the review of https://github.com/opf/openproject/pull/3196, a
notification box with error style was used to communicate a warning
prior to repository deletion with *severe* consequences
(removal of files outside OpenProject).

This commit suggests to add a `-severe` switch to warning notifications
to signal such a severe decision while NOT being an error.

The implementation re-uses the error colors (as red is often used
to communicate severe warnings - cf. GitHub danger zone, Bootstrap
styles).

I envision this notification to be used in other aspects of the
application, with at least project destroy_info to be of the same
severity that I argue this to be included in the core.

Relevant work package: https://community.openproject.org/work_packages/20999