Use severe notification for project destroy dialog
The current project `destroy_info` style is inconsistent with the
styleguide, as it uses a `notification` (used for one-line notifications
only) instead of a warning notification box.

This commit switches this dialog to the notification-box with warning
parameter and suggests to additionaly use the new `-severe` switch for warnings
due the potential ramifications this action has.

Using `-severe` also makes it compatible with the upcoming repository
destroy_info dialog, which also uses this dialog when local data is
to be deleted outside OpenProject.