Remove delayed running of management workers.
This commit (temporarily) removes the delayed creation and deletion of
managed repositories for the sake of improvded error handling.
After consultation with @ulferts, we suggest to improve the `Delayed::Job`
pipeline to create means for identifying jobs (referenced ID) and to
create a unified service to manage database AND filesystem access in a
transactional manner.

Also provides the following corrections:

- Improve migration speed
- Delete parent projects only by their identifier, not the managed root