Change feature spec broken by openproject-checkout.
The checkout plugin overrides behavior in the core which leads
to the repository settings form submit button to be displayed
even though the core feature expects it to be hidden.

As the surrounding spec is relevant for the plugin, too, and should not
be skipped / disabled, this commit instead reworks the test to check
for a different aspect of the page to ensure the request was successful.