explicitly define order of groups in PSQL
Always remember kids: SQL databases do not guarantee an order if you don't ask for one!

That is why PostgreSQL sometimes found a predecessor for an aggregated journal that was not supposed to have one: Based on internal decisions the order of rough groups was not ascending (as expected implicitly) and master and addition would switch places.

ALWAYS SPECIFY AN ORDER IF YOU EXPECT ONE!