Use find selector after JS redirect
On occasions, the CI reports the same feature spec to be failing
due to a StaleElementReferenceError, however the stacktrace suggests
that this error stems from the extraction of node texts after the actual
query within one retry of the synchronized selection.

The stacktrace begins that error inside `synchronized` and deep within
`result.failure_message`.
https://github.com/jnicklas/capybara/blob/120daaa8563d8b06f4727580ab58cd04cb5856aa/lib/capybara/node/matchers.rb#L97

Thus this commit suggests to use the simplest `find` selector to locate
the element, which does not access its contents.