Disallow setting the owner of a repository folder
The original repository management epic allows to define
an owner that a created repository is set to.

As this will not work for non-root users on *nix systems,
this commit removes this functionality in favor of setting
only the group.

This is allowed by non-privileged users _as long as_ users are
member of this group.