prevent too long Request-URLs for many watchers
turn back to the older behaviour of only sending checked checkboxes,
which is favorable when there are many watchers available.
Otherwise changing the type (which issues a GET request) creates too long URLs for many watchers.

A scalable alternative could be to issue a POST request, but we only have so much time and the legacy view is already deprecated...