fix sprockets/sprockets-rails dependency
The 2_2_2_backport2 at https://github.com/tessi/sprockets/tree/2_2_2_backport2 actually has the version 2.2.3 leading to an invalid dependency with sprockets-rails.
This fix relies on a clone of the repo to fix the branch structure. Note, that this is discontinued in OP 5.0 where we again rely on the standard sprockets.