Delete repository: Fix internal error on lacking permissions
While rolling out on dev, we found that apache by default modifies
the access permissions of a repository.
This in turn causes the deletion of a repository to no longer complete
sucessfully, but return with a permission error.

This error was not caught an escalated into a 500 on the repository
settings.