Use default project type when none available
When the project has no types available, and work package is created
through the factory, it will raise an inclusion error from
`work_package/validations`.

This change forces projects to have a default type in the case
they do not (yet).

This also fixes a wrong assignment of type from name, which however
doesn't seem to be used (yet).