Work around outdated group SQL
The `Group.includes(:users)` call from the GroupsController seems
to return an outdated / incomplete result of the members.

You can verify that by calling `@group.reload` after the affected line,
which will yield the correct result.

I assume that this has something to do with the HABTM association
`:users` for which a similar issue appears to be still open.

Using `eager_loading` directly seems to fix the issue, however
we have to remove :memberships here as it requires a condition.