Storage performance and accuracy improvements
This commit allows storage counting to shell out to `du` where
available.

The method uses du with `-bs` to count the apparent file size in bytes.
Note that apparent size is NOT the same as the real required size on
file system, but is used to be consistent with other sizes displayed in
OpenProject (e.g., attachments size)

When du is either unavailable or incompatible, in that it doesn't
return the requested count, the ruby fallback counting method is used.

This method has been slightly altered to use `File.stat` instead of
`File.new(path).size` for reduced computational overhead compared to
`du`.