Refuse to delete a repository in OpenProject on FS permission errors
This commit changes the behavior of removing repositories.
It is no longer possible to remove a managed repository (model) when the
underlying repository directory could not be removed.

This error originally occurred when Apache committed changes to the repository
through mod_svn, which resulted in the OpenProject user being unable to
delete a particular subdirectory.