Enforce passing of current_user into representers
Thereby permission checks will no longer silently fail if no user was
passed.
This uncovered a lot of places where we actually expected a user to be passed, when it wasn't.
This was especially true for collections.
The basic idea is, that virtually all representers could change their behaviour based on the current user.
Representers where this is not the case (e.g. Digest), can still explicitly opt-out.